Rely only on connection string for failover partners
Remove the partnersCache, and instead rely only on the failoverpartner
value from the connection string. This simplifies the code and fixes
issues with creating new connections to the database where the principal
was found at failoverpartner (e.g. An initial connection could not be
made to the instance specified in the server parameter).