Use interface for logging
Previously the driver would only accept setting the logger to stdlib
*log.Logger structs. This change allows anything with a logger-like
interface to be used instead.

In order to make logging work even when a logger isn't explicitly
specified, an optionalLogger struct has been added that still works (by
logging to the global stdlib logger) as its zero value. It fills the
same function as the Logger struct it's replacing, though the naming
makes things a little more obvious.

Another option I explored was just using the new Logger interface
directly as the type of log in MssqlDriver. We could provide a default
struct that implements Logger and just forwards to the global stdlib log
functions

The problem was that this default needed to be provided everywhere a
MssqlDriver is created. Since log is private, that meant that no one
outside of the package could ever create an instance of the public
MssqlDriver that would actually work. Honestly I don't know that they
ever _would_, or whether it could be alleviated by providing a public
constructor, but in the interest of keeping things compatible I opted
for the optionalLogger strategy I outlined above.