add UniqueIdentifier type
SQL Server's UniqueIdentifier type uses native endianness for some of
its fields within the UUID. Add a new type, UniqueIdentifier, that
correctly handles scanning based on the type of data being scanned.

Scanning a []byte swaps the bytes of the fields that SQL Server stores
using native endianness.

Scanning a string decodes the hex-encoded values without swapping any
bytes.

Implement fmt.Stringer so that the returned value of String is congruent
with the behavior of CAST(myuniqueidentifer AS varchar(36)).