change underlying type of UniqueIdentifier
Change the underlying type of UniqueIdentifier to ([]byte -> [16]byte).

Changing the type to an array also allowed UniqueIdentifier.Equal to be
removed since array types are comparable.

Modify UniqueIdentifier.String to return a string that uses capitalized
hex encoding for closer parity with SQL Server's conversion of its
UniqueIdentifer type to character types.