Reduce tds buffer size to match packetsize
We noticed that when connecting to many databases, the amount of memory being utilized by the newTdsBuffer() func was rather significant, especially in comparison to the business logic that utilized the connections. I was able to use pprof to figure out that the newTdsBuffer() func was the source of the memory utilization and that these 2 buffers get initialized to a fairly large size. This reduces the size of the buffers to match the packetsize, significantly reducing the amount of memory used for the buffers in the majority use case.