feat: add Always Encrypted support
This commit adds partial support for the Microsoft SQL
"Always Encrypted" feature (basically, E2E encryption).

The current implementation is to be consider a __preview__
since it might not be perfectly implemented.

Supported features:
- PFX "keystore"
- Seamless encryption

Missing features:
- Support for Private Keys that are not RSA
- Encryption support (only Decryption is possible at the moment)

The most probably needs to be improved a bit, but so far it's
working for some of the use cases that I needed it for.
Feel free to test the feature and open an issue if you find any
problem: my goal is to have enough testers to spot eventual bugs.

fix: lint issues