Export auth interface to allow external implementations e.g. krb5 - V1-Candidate (#15)
* Export auth interface to allow external implementations

The original driver auth interface and implementations have been moved into their own packages and exported. There are two implementations within the driver, NTML and Windows kerberos.
As before the default on windows is the winsspi package and on, on Linux, NTLM.

The consuming application can now override this at runtime by calling mssql.SetAuthProvider(authProvider)

This allows the application to provide a custom implementation of the auth.Provider interface.
auth.Provider is in turn responsible for creating an instance of auth.Auth for the required authentication mechanism. windows, NTML etc.

* remove unused function and merge master changes

* add golang.org/x/crypto/md4 to appveyor.yml

* add unit tests for SetAuthProvider and getAuth

* Rename Auth to IntegratedAuthenticator

* change package references

* modular authentication and runtime registration

* fix windows sspi

* change test to use DeepEqual instead of struct equality

* whitespace change to assess appveyor issue

* fix linting issues

* bring the kerberos authenticator into the driver

* Revert "bring the kerberos authenticator into the driver"

This reverts commit 64370775106424e87dd7c277d981b12332c139c0.